requirements.txt
streamlit==1.22.0
tensorflow==2.11.0
pillow==9.2.0
numpy==1.21.5
keras==2.11.0
